fix: resolve data fetching by category and search for listings
Fixed issues with fetching housing data based on category and search input.
Additionally, corrected the display of success messages for profile
creation and housing listing creation.
