+++ components/mainPages/staysPage/StaysList.tsx
@@ -1,21 +1,4 @@
-import { fetchProperties } from '@/utils/actions';
-import EmptyList from '../common/EmptyList';
-
-type StaysListProps = {
-	category?: string;
-	search?: string;
-}
-
-async function StaysList(props: StaysListProps) {
-	const { category, search } = props;
-	const properties = await fetchProperties({});
-
-	if (properties.length === 0) {
-		return (
-			<EmptyList />
-		);
-	}
-
+function StaysList() {
 	return (
 		<div>
 			Список жилья
