+++ components/mainPages/staysPage/index.ts
@@ -1,2 +1,2 @@
 export { default as StaysCategoriesList } from './StaysCategoriesList';
-export { default as StaysList } from './StaysList';
+export { default as StaysContainer } from './StaysContainer';
